✨ feat: Add new env variables to setup-and-load-env action
This commit updates the `setup-and-load-env` composite action to
accept more inputs corresponding to environment variables defined in
`.env.production`, and modifies the CD workflow to pass these secrets
as inputs to the action.

This commit adds several new environment variables as secrets to the
GitHub workflow, and ensures these variables are available to the
deployment script by explicitly exporting them. This facilitates
configuring the application with sensitive information required at
runtime.

This commit updates the GitHub Actions workflow to include more
environment variables as secrets, such as those related to admin
credentials, Twilio, and VPS configuration, and bumps the package
version from 0.0.7 to 0.0.8.