✨ feat: Refactor DB schema to use TradesMan ID & cascade deletes
This commit refactors the database schema to use `TradesMan` ID
instead of `User` ID in related tables like `Docs`,
`BusinessDetail`, `ServiceArea`, and `PaymentMethod`.  It also
implements cascade deletes to ensure data consistency when a
`TradesMan` is deleted.  Additionally, it updates the admin
service to correctly delete users, including associated
`TradesMan` data, and increments the number of interested and
shortlisted users in the job activity when proposals are created or
accepted.